extract mock-gc.js out from test-liveslots.js
This utility creates manually-controlled GC utilities: WeakRef and
FinalizationRegistry
